Fixed a problem in the RI found during testing with the OSGi Security Manager enabled.
Previously, the RI used the JNDI caller's BundleContext to track JNDI services on behalf of the client.  When the Security Manager is enabled, this approach causes problems, since the OSGi framework will not report service lifecycle events to Bundles that do not contain the correct "get" permission (in this case, for the "javax.naming.spi.*" services).  

This change fixes this problem by using the JNDI implementation's BundleContext to track JNDI provider services, while still using the JNDI client's BundleContext to obtain the service on behalf of the client.  This makes it possible to guarantee class-space consistency while still receiving the proper service lifecycle events when the Security Manager is enabled.